bluetooth: id: fix uninitialized RPA for non-default identities
This change fixes uninitialized RPA values for advertising sets of
non-default Bluetooth identities. The regression has been introduced
in the following commit:

https://github.com/zephyrproject-rtos/zephyr/commit/8d6b206064c915316245616a718ae29eb157a717

The follow-up fix for the default Bluetooth identity has been
introduced in a separate commit:

https://github.com/zephyrproject-rtos/zephyr/commit/88c20b9cddf0bd5e5d6c163611b2563f9266be6f

Non-default Bluetooth identities can be loaded from the Settings
subsystem. In this case, RPAs of related advertising sets to these
identities are left uninitialized as the bt_id_create() function
context is not executed. As a result, the RPA is not created for
advertising sets of non-default Bluetooth identities, and the
advertising is started with the 00:00:00:00:00:00 address.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>
(cherry picked from commit 29a67338d12a298baeb85c68a53d255e1ddf2fa3)